Next sgrenier

+++ gameplay-samples/sample00-mesh/src/MeshGame.h
@@ -47,12 +47,12 @@ protected:
     /**
      * @see Game::update
      */
-    void update(long elapsedTime);
+    void update(float elapsedTime);
 
     /**
      * @see Game::render
      */
-    void render(long elapsedTime);
+    void render(float elapsedTime);
 
 private:
